Refactor AccountModule
